Recogniton-AI/Face-Recognition/requirements/requirements_pip.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1615411197583/work
astor==0.8.1
certifi==2020.12.5
chardet==4.0.0
coverage @ file:///tmp/build/80754af9/coverage_1614614208500/work
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1618435161347/work
facenet==1.0.5
faiss-cpu==1.7.0
gast==0.2.2
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1614884176509/work
h5py @ file:///tmp/build/80754af9/h5py_1593454121459/work
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617877310517/work
imutils==0.5.4
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
Keras==2.3.1
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.3.1
Markdown @ file:///tmp/build/80754af9/markdown_1614363833670/work
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
opt-einsum==3.1.0
Pillow==8.2.0
protobuf==3.14.0
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1614446678517/work
scipy==1.1.0
six @ file:///tmp/build/80754af9/six_1605205335545/work
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
